fetch new checkins

+++ tools/fetch-swarm-checkins.ts
@@ -1,30 +1,18 @@
-import { writeFileSync, readdirSync, readFileSync } from 'fs';
+import { writeFileSync, readFileSync, existsSync } from 'fs';
 import axios from 'axios';
 import { Checkin, Root } from './swarm';
+import { mapGen, readAllJson } from './data';
 
 function readCreds() {
   return JSON.parse(readFileSync('creds/swarm.json', 'utf8'));
 }
 
-function findEarliestCheckin() {
-  const files = readdirSync('data/swarm');
-  const timestamps = files.map((filename) => {
-    const checkin: Checkin = JSON.parse(
-      readFileSync(`data/swarm/${filename}`, 'utf8')
-    );
-    return checkin.createdAt;
-  });
-
-  if (timestamps.length === 0) {
-    return undefined;
-  }
-
-  return Math.min(...timestamps);
-}
-
-export default async function () {
+export default async function ({ _: [direction = 'new'] }: { _: string[] }) {
+  const checkins = readAllJson<Checkin>('swarm');
+  const timestamps = mapGen(checkins, (c) => c.createdAt);
   // assuming running the script repeatedly after the API returns a few 500s
-  let lastTs: undefined | number = findEarliestCheckin();
+  let lastTs: undefined | number =
+    direction === 'new' ? Math.max(...timestamps) : Math.min(...timestamps);
 
   do {
     const params = new URLSearchParams({
@@ -32,8 +20,11 @@ export default async function () {
       v: '20200101',
       limit: '100',
     });
-    if (lastTs !== undefined) {
-      params.append('beforeTimestamp', lastTs.toString());
+    if (isFinite(lastTs)) {
+      params.append(
+        direction === 'new' ? 'afterTimestamp' : 'beforeTimestamp',
+        lastTs.toString()
+      );
     }
 
     const url = `https://api.foursquare.com/v2/users/self/checkins?${params.toString()}`;
@@ -52,6 +43,11 @@ export default async function () {
     lastTs = undefined;
     for (const checkin of response.response.checkins.items) {
       const filename = `data/swarm/${checkin.id}.json`;
+      if (existsSync(filename)) {
+        lastTs = undefined;
+        break;
+      }
+      console.log(filename);
       writeFileSync(filename, JSON.stringify(checkin, null, 2));
       lastTs = checkin.createdAt;
     }
